add a script to merge records of multiple databases
The timetagger-mergedatabases.py is intended to merge
TimeTagger record tables of multiple user databases
into one user database, replacing the existing records.
Extends the records by user information.

This can be used as a workaround to create a unified view
of the records of all or a group of users.